resolve reverse dependency order before service remove
